refactor: Move existing SchemaValidator functionality into its own module
To enable reuse by the schema service for schema upsert in a future PR.
